Move GameSettings and DefaultSettings into game_settings module
